v3 Rewrite - Tuesday
Fixed backup.js mc commands prefixed with a /
Split util into multiple modules to remove bloat for imports.
Migrated  some core functions into modul from util.
Changed discordStdin to use new xMP structure.
Restored accidentally deleted update for serverSettings.json
Added checks for server process existing before killing it on exit.
Added new argument for server start to  help identify it in htop.
Began moving log.js functions into the modules alongside the functions they belong to.
Rewrote math.js to use new xMP structure and handle logging locally.

+++ ForgeTest/serverWrapper.js
@@ -29,7 +29,7 @@ let server = null;
 let sSFile = './serverSettings.json'
 let sS = require(sSFile);
 let loadedModules = {};
-let serverStartVars = Object.assign([], sS.serverStartVars);
+let serverStartVars = Object.assign([], [`-D${sS.serverName}`].concat(sS.serverStartVars));
 serverStartVars.push("-Xms"+sS.minRamAllocation, "-Xmx"+sS.maxRamAllocation, "-jar", sS.jar)
 serverStartVars = serverStartVars.concat(sS.serverPostfixVars);
 sS.server_dir = __dirname;
@@ -43,7 +43,7 @@ if (((sS.modules['discord']||{}).settings||{}).discord_token == "") {
 }
 
 const cleanExit = () => {
-	server.process.kill();
+	if (server && server.process) server.process.kill();
 	process.kill();
 };
 process.on('SIGINT', cleanExit); // catch ctrl-c
